+++ src/config.h
@@ -59,6 +59,7 @@ typedef struct ns_config {
     gboolean speculative_preload;
     gboolean async_image_decode;
     gboolean images_enabled;
+    gboolean javascript_enabled;
     gboolean webgl_enabled;
     gboolean camera_enabled;
     gboolean microphone_enabled;